move directory, update integration test
