+++ upstash/kafka/credential/api_calls.go
@@ -19,7 +19,6 @@ func deleteCredential(c *client.UpstashClient, credentialId string) (err error) 
 	return c.SendDeleteRequest("/v2/kafka/credential/"+credentialId, nil, "Delete Kafka Credential")
 }
 
-//
 func getCredential(c *client.UpstashClient, credentialId string) (credential Credential, err error) {
 	resp, err := c.SendGetRequest("/v2/kafka/credentials", "Get Kafka Cluster")
 	var credentials []Credential
